# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/couriers": {
            "get": {
                "tags": [
                    "Couriers"
                ],
                "summary": "Get all couriers",
                "operationId": "0431baf1d6489ad0059e7efbbeee9c58",
                "responses": {
                    "200": {
                        "description": "Couriers retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Couriers"
                ],
                "summary": "Create a new courier",
                "operationId": "1a23eb41d68e521fdb65083bd2dfd55d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Name"
                                ],
                                "properties": {
                                    "Name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "LicenseNumber": {
                                        "type": "string",
                                        "example": "B1234XYZ"
                                    },
                                    "VehicleType": {
                                        "type": "string",
                                        "example": "Car"
                                    },
                                    "PhoneNumber": {
                                        "type": "string",
                                        "example": "082233445566"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Courier created successfully"
                    }
                }
            }
        },
        "/couriers/{id}": {
            "put": {
                "tags": [
                    "Couriers"
                ],
                "summary": "Update a courier",
                "operationId": "f2f21f7f912d6566b0c77a1f507eebda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Courier ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Name": {
                                        "type": "string",
                                        "example": "Jane Updated"
                                    },
                                    "LicenseNumber": {
                                        "type": "string",
                                        "example": "D4321CBA"
                                    },
                                    "VehicleType": {
                                        "type": "string",
                                        "example": "Truck"
                                    },
                                    "PhoneNumber": {
                                        "type": "string",
                                        "example": "085566778899"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Courier updated successfully"
                    },
                    "404": {
                        "description": "Courier not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Couriers"
                ],
                "summary": "Delete a courier",
                "operationId": "ac932b63e8aa13d5d60d65634f421f8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Courier ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Courier deleted successfully"
                    },
                    "404": {
                        "description": "Courier not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}